+++ src/components/login/LoginForm/EmailInput/index.tsx
@@ -1,0 +1,5 @@
+import { emailInputStyle } from "./style";
+
+export default function EmailInput() {
+  return <input type="email" placeholder="이메일" css={emailInputStyle} />;
+}
